Создает модули из скрипта
